alter procedure if wifi lost
Found that wifi was being dropped more often (and for longer periods) than was ideal. Original code caused the bot to disable when wifi was lost, so that it wouldn't become a runaway. However, this meant the bot toppled over too frequently. These changes just cause the bot to stop driving while the wifi is lost -- this prevents it from becoming a runaway, but doesn't cause it to fall.